+++ bundles/org.openhab.binding.surepetcare/src/main/java/org/openhab/binding/surepetcare/internal/dto/SurePetcarePhoto.java
@@ -30,5 +30,5 @@ public class SurePetcarePhoto extends SurePetcareBaseObject {
     // }
 
     public String location;
-    public Integer uploadingUserId;
+    public Long uploadingUserId;
 }
